backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
# --index-url https://pypi.tuna.tsinghua.edu.cn/simple
# --trusted-host pypi.tuna.tsinghua.edu.cn

aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.7
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
alembic==1.13.3
    # via -r requirements.in
amqp==5.3.1
    # via kombu
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anthropic==0.39.0
    # via -r requirements.in
anyio==4.12.1
    # via
    #   anthropic
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via -r requirements.in
attrs==25.4.0
    # via aiohttp
bcrypt==4.2.1
    # via
    #   -r requirements.in
    #   passlib
billiard==4.2.4
    # via celery
celery==5.4.0
    # via
    #   -r requirements.in
    #   flower
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
cryptography==43.0.3
    # via
    #   -r requirements.in
    #   pyjwt
    #   python-jose
deprecated==1.3.1
    # via
    #   limits
    #   pygithub
distro==1.9.0
    # via
    #   anthropic
    #   openai
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.2.0
    # via -r requirements.in
esprima==4.0.1
    # via -r requirements.in
fastapi==0.128.0
    # via -r requirements.in
flower==2.0.1
    # via -r requirements.in
gitpython==3.1.43
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.27.2
    # via
    #   -r requirements.in
    #   anthropic
    #   openai
humanize==4.15.0
    # via flower
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
jiter==0.12.0
    # via
    #   anthropic
    #   openai
kombu==5.6.2
    # via celery
limits==5.6.0
    # via slowapi
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
neo4j==5.24.0
    # via -r requirements.in
networkx==3.4.2
    # via -r requirements.in
openai==1.54.5
    # via -r requirements.in
packaging==25.0
    # via
    #   kombu
    #   limits
passlib[bcrypt]==1.7.4
    # via -r requirements.in
prometheus-client==0.21.0
    # via
    #   -r requirements.in
    #   flower
prompt-toolkit==3.0.52
    # via click-repl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psycopg[binary]==3.2.13
    # via -r requirements.in
psycopg-binary==3.2.13
    # via psycopg
pyasn1==0.6.2
    # via
    #   python-jose
    #   rsa
pycparser==2.23
    # via cffi
pydantic==2.10.5
    # via
    #   -r requirements.in
    #   anthropic
    #   fastapi
    #   openai
    #   pydantic-settings
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pygithub==2.5.0
    # via -r requirements.in
pyjwt[crypto]==2.9.0
    # via
    #   -r requirements.in
    #   pygithub
pynacl==1.6.2
    # via pygithub
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   celery
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.3.0
    # via -r requirements.in
python-json-logger==3.2.1
    # via -r requirements.in
watchtower==3.3.1
    # via -r requirements.in
boto3==1.35.94
    # via
    #   -r requirements.in
    #   watchtower
botocore==1.35.94
    # via
    #   boto3
    #   watchtower
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
s3transfer==0.10.4
    # via boto3
python-multipart==0.0.12
    # via -r requirements.in
pytz==2024.2
    # via
    #   -r requirements.in
    #   flower
    #   neo4j
pyyaml==6.0.3
    # via uvicorn
redis==5.2.0
    # via -r requirements.in
requests==2.32.3
    # via pygithub
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
slowapi==0.1.9
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anthropic
    #   httpx
    #   openai
sqlalchemy==2.0.35
    # via
    #   -r requirements.in
    #   alembic
starlette==0.50.0
    # via fastapi
tornado==6.5.4
    # via flower
tqdm==4.67.1
    # via openai
tree-sitter==0.23.2
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   alembic
    #   anthropic
    #   fastapi
    #   limits
    #   openai
    #   pydantic
    #   pydantic-core
    #   pygithub
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via
    #   celery
    #   kombu
    #   psycopg
urllib3==2.6.3
    # via
    #   pygithub
    #   requests
uvicorn[standard]==0.32.0
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.2.14
    # via prompt-toolkit
websockets==13.1
    # via
    #   -r requirements.in
    #   uvicorn
wrapt==1.16.0
    # via deprecated
yarl==1.22.0
    # via aiohttp
tenacity==9.0.0
    # via -r requirements.in
aiofiles==24.1.0
    # via -r requirements.in
opentelemetry-api==1.28.2
    # via -r requirements.in
opentelemetry-sdk==1.28.2
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-instrumentation
opentelemetry-semantic-conventions==0.49b2
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
opentelemetry-instrumentation==0.49b2
    # via
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-redis
    #   opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-asgi==0.49b2
    # via opentelemetry-instrumentation-fastapi
opentelemetry-util-http==0.49b2
    # via
    #   opentelemetry-instrumentation-asgi
    #   opentelemetry-instrumentation-fastapi
    #   opentelemetry-instrumentation-httpx
grpcio==1.69.0
    # via opentelemetry-exporter-otlp-proto-grpc
protobuf==5.29.3
    # via
    #   grpcio
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-common==1.28.2
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-grpc==1.28.2
    # via -r requirements.in
opentelemetry-proto==1.28.2
    # via
    #   opentelemetry-exporter-otlp-proto-common
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi==0.49b2
    # via -r requirements.in
opentelemetry-instrumentation-httpx==0.49b2
    # via -r requirements.in
opentelemetry-instrumentation-redis==0.49b2
    # via -r requirements.in
opentelemetry-instrumentation-sqlalchemy==0.49b2
    # via -r requirements.in
